+++ Release_Notes.md
@@ -6,6 +6,7 @@
 ### Improvements
 - Removed centralManagement from the BClientChoices constructor ... it's not needed
 - Performance Improvement - Added a BusinessLogic property to the BClientChoices to limit how many times the a BClientChoices is created
+- DirectoryUtility: Changed getBusinessLogic from a method to a property and renamed it to BusinessLogic
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be causing a logout to occure